requirements.txt
streamlit
pandas
langchain
langchain-openai
langchain-community
python-dotenv
